Refactor form validation into `ValidatingForm`
- Create `ValidatingForm` component to check form validity only upon
  attempting submission and display validation feedback as needed
