+++ src/components/index.js
@@ -1,2 +1,3 @@
 export { default as Footer } from "./Footer/Footer";
 export { default as Navigation } from "./Navigation/Navigation";
+export { default as ValidatingForm } from "./ValidatingForm/ValidatingForm";
